Revert "Merge pull request #33638 from coolljt0725/mnt_deatch"
This reverts commit ff2127968dd77be6a0d5d36e28a3ef465a143e15, reversing
changes made to f4bf61c79da153c039e2df586737cd8d54d26a55.

The reverted patch was intentionally making it more likely that a
reference will be left behind on umount and makes mount / umount racy.

This is not helped at all by the fact that docker does mount/umount
pairs twice in the hotpath for a container setup before the final mount.

Related to discussion in https://github.com/moby/moby/issues/34672
Reverting this patch by itself doesn't fix that issue, but it logically
should make that sort of issue worse.

On kernels prior to v4.13, this detached mnt doesn't make things worse,
but with the inclusion of https://patchwork.kernel.org/patch/9760547/,
it makes the change logically incorrect.
That patch also makes it clear upstream considers remounting a
still-in-use overlayfs mount a bug, which the reverted commit made
logically more likely to happen.

Signed-off-by: Euan Kemp <euan.kemp@coreos.com>